requirements.txt
Flask>=3.0.0
python-dotenv>=1.0.0
dnspython>=2.6.1
requests>=2.31.0
tqdm>=4.66.0